+++ src/test/helpers/bitbucketNock.ts
@@ -1,6 +1,12 @@
 /* eslint-disable @typescript-eslint/camelcase */
 import nock from 'nock';
 import { BitbucketCommit } from '../../services/bitbucket/BitbucketService';
+import { BitbucketPullRequestState } from '../../services/git/IVCSService';
+import qs from 'qs';
+import { Paginated } from '../../inspectors';
+import { PullRequest } from '../../services/git/model';
+import { getPullRequestResponse } from '../../services/git/__MOCKS__/bitbucketServiceMockFolder';
+import _ from 'lodash';
 
 export class BitbucketNock {
   user: string;
@@ -12,22 +18,31 @@ export class BitbucketNock {
     this.url = 'https://api.bitbucket.org/2.0';
   }
 
-  getApiResponse(resource: string, id?: number | string, value?: string): nock.Scope {
+  getApiResponse(
+    resource: string,
+    id?: number | string,
+    value?: string,
+    state?: BitbucketPullRequestState | BitbucketPullRequestState[],
+  ): nock.Scope {
     let url = `${this.url}/repositories/${this.user}/${this.repoName}/${resource}`;
     let response;
 
-    const params = {};
+    let params = {};
     const persist = true;
+
+    if (state === undefined) {
+      state = BitbucketPullRequestState.open;
+    }
 
     if (value !== undefined) {
       switch (value) {
         case 'comments':
           url = url.concat(`/${id}/${value}`);
-          response = new IssueComments().issueComments;
+          response = new IssueCommentsMock().issueComments;
           break;
         case 'commits':
           url = url.concat(`/${id}/${value}`);
-          response = new Commits().commits;
+          response = new CommitsMock().commits;
           break;
       }
     } else {
@@ -35,25 +50,44 @@ export class BitbucketNock {
         case 'pullrequests':
           if (id !== undefined) {
             url = url.concat(`/${id}`);
-            response = new PullRequest().pullRequest;
+            response = new PullRequestMock(<BitbucketPullRequestState>state).pullRequest;
           } else {
-            response = new PullRequests().pullrequests;
+            if (state === BitbucketPullRequestState.open) {
+              const pullRequest = new PullRequestMock(state).pullRequest;
+              response = new PullRequestsMock([pullRequest]).pullrequests;
+            } else {
+              const stateForUri = qs.stringify({ state: state }, { addQueryPrefix: true, indices: false, arrayFormat: 'repeat' });
+              url = url.concat(`${stateForUri}`);
+              params = { state: state };
+
+              if (typeof state !== 'string') {
+                const pullRequests: Bitbucket.Schema.Pullrequest[] = [];
+                state.forEach((state) => {
+                  pullRequests.push(new PullRequestMock(state).pullRequest);
+                });
+
+                response = new PullRequestsMock(pullRequests).pullrequests;
+              } else {
+                const pullRequest = new PullRequestMock(<BitbucketPullRequestState>state).pullRequest;
+                response = new PullRequestsMock([pullRequest]).pullrequests;
+              }
+            }
           }
           break;
         case 'issues':
           if (id !== undefined) {
             url = url.concat(`/${id}`);
-            response = new Issue().issue;
+            response = new IssueMock().issue;
           } else {
-            response = new Issues().issues;
+            response = new IssuesMock().issues;
           }
           break;
         case 'commits':
-          response = new RepoCommits().repoCommits;
+          response = new RepoCommitsMock().repoCommits;
           break;
         case 'commit':
           url = url.concat(`/${id}`);
-          response = new RepoCommit().repoCommit;
+          response = new RepoCommitMock().repoCommit;
           break;
         default:
           throw Error('You passed wrong value or id');
@@ -77,9 +111,35 @@ export class BitbucketNock {
     }
     return interceptor;
   }
+
+  mockBitbucketPullRequestsResponse(states: BitbucketPullRequestState | BitbucketPullRequestState[]): Paginated<PullRequest> {
+    const pullRequests: PullRequest[] = [];
+
+    const paginatedPullrequests: Paginated<PullRequest> = {
+      items: [getPullRequestResponse],
+      hasNextPage: true,
+      hasPreviousPage: false,
+      page: 1,
+      perPage: typeof states === 'string' ? 1 : states.length,
+      totalCount: typeof states === 'string' ? 1 : states.length,
+    };
+
+    if (typeof states !== 'string') {
+      states.forEach((state) => {
+        const pullrequest = _.cloneDeep(getPullRequestResponse);
+        pullrequest.state = state;
+        pullRequests.push(pullrequest);
+      });
+      paginatedPullrequests.items = pullRequests;
+    } else {
+      getPullRequestResponse.state = states;
+      paginatedPullrequests.items = [getPullRequestResponse];
+    }
+    return paginatedPullrequests;
+  }
 }
 
-export class Issues {
+export class IssuesMock {
   issues: Bitbucket.Schema.PaginatedIssues;
   constructor() {
     this.issues = {
@@ -87,12 +147,12 @@ export class Issues {
       size: 3070,
       page: 1,
       next: 'https://api.bitbucket.org/2.0/repositories/pypy/pypy/issues?page=2',
-      values: [new Issue().issue],
+      values: [new IssueMock().issue],
     };
   }
 }
 
-export class Issue {
+export class IssueMock {
   issue: Bitbucket.Schema.Issue;
   constructor() {
     this.issue = {
@@ -192,23 +252,25 @@ export class Issue {
   }
 }
 
-export class PullRequests {
+export class PullRequestsMock {
   pullrequests: Bitbucket.Schema.PaginatedPullrequests;
 
-  constructor() {
+  constructor(pullRequests: (Bitbucket.Schema.Pullrequest & Bitbucket.Schema.Object)[]) {
     this.pullrequests = {
       page: 1,
       next: 'https://api.bitbucket.org/2.0/repositories/pypy/pypy/pullrequests?page=2',
       pagelen: 10,
       size: 20,
-      values: [new PullRequest().pullRequest],
+      values: pullRequests,
     };
   }
 }
 
-export class PullRequest {
+export class PullRequestMock {
   pullRequest: Bitbucket.Schema.Pullrequest & Bitbucket.Schema.Object;
-  constructor() {
+  state: BitbucketPullRequestState;
+  constructor(state: BitbucketPullRequestState) {
+    this.state = state;
     this.pullRequest = {
       description: 'Added a floor() ufunc to micronumpy',
       links: {
@@ -299,7 +361,7 @@ export class PullRequest {
         },
       },
       comment_count: 0,
-      state: 'DECLINED',
+      state: this.state,
       task_count: 0,
       participants: [],
       reason: 'already merged',
@@ -349,7 +411,7 @@ export class PullRequest {
   }
 }
 
-export class Commits {
+export class CommitsMock {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   commits: any;
   constructor() {
@@ -502,7 +564,7 @@ export class Commits {
   }
 }
 
-export class IssueComment {
+export class IssueCommentMock {
   issueComment: Bitbucket.Schema.IssueComment;
   constructor() {
     this.issueComment = {
@@ -573,19 +635,19 @@ export class IssueComment {
   }
 }
 
-export class IssueComments {
+export class IssueCommentsMock {
   issueComments: Bitbucket.Schema.PaginatedIssueComments;
   constructor() {
     this.issueComments = {
       pagelen: 20,
-      values: [new IssueComment().issueComment],
+      values: [new IssueCommentMock().issueComment],
       page: 1,
       size: 14,
     };
   }
 }
 
-export class RepoCommit {
+export class RepoCommitMock {
   repoCommit: Bitbucket.Schema.Commit;
   constructor() {
     this.repoCommit = {
@@ -692,11 +754,11 @@ export class RepoCommit {
   }
 }
 
-export class RepoCommits {
+export class RepoCommitsMock {
   repoCommits: BitbucketCommit;
   constructor() {
     this.repoCommits = {
-      values: [new RepoCommit().repoCommit],
+      values: [new RepoCommitMock().repoCommit],
       pagelen: 20,
       page: 1,
       size: 14,